+++ docs/Changelog.md
@@ -6,6 +6,7 @@ v0.9.67 Oct 1, 2025
 + New: Added a Validate Port option to the Local Port context menu, which checks whether the port is valid and unique, and proposes a new port if it is not.
 + New: Added a Propose New Port option to the Local Port context menu, which finds an available unique port and prompts you to apply it.
 + New: Added Validate Port and Propose New Port options to the More Tunnels table context menu, providing the same functionality as in the Local Port context menu.
++ New: Jump search extended to include local tunnel port numbers.
 
 v0.9.66 Sep 29, 2025
 ------------------